Add HIR and MIR Expression Helpers for Enhanced Expression Handling
- Introduced new helper functions in `lower_expr_helpers.go` and `lower_expr_tag.go` to streamline expression handling in HIR and MIR.
- Implemented functions for looking up strings, creating variable references, and managing reference types, improving code organization and readability.
- Added intrinsic function handling in `intrinsic_helpers.go` and `intrinsic_io.go`, enhancing the VM's capabilities for managing intrinsic operations.
- Refactored existing functions in `lower_expr.go` to utilize the new helpers, reducing code duplication and improving maintainability.
- These changes aim to enhance the robustness and clarity of expression handling across the codebase.
