+++ STATS.md
@@ -4,27 +4,27 @@
 
 ## 📊 Main code (without tests)
 
-- **Files:** 392
-- **Lines of code:** 79357
+- **Files:** 400
+- **Lines of code:** 79602
 
 ## 📁 Directory breakdown
 
 | Directory | Files | Lines |
 |------------|--------|-------|
 | `cmd/` | 12 | 2355 |
-| `internal/` | 380 | 77002 |
+| `internal/` | 388 | 77247 |
 
 ## 🏆 Top 10 packages by size
 
 | # | Package | Lines |
 |---|-------|-------|
 | 1 | `internal/sema` | 19042 |
-| 2 | `internal/vm` | 8437 |
+| 2 | `internal/vm` | 8558 |
 | 3 | `internal/parser` | 7976 |
-| 4 | `internal/hir` | 5729 |
+| 4 | `internal/hir` | 5753 |
 | 5 | `internal/driver` | 4620 |
 | 6 | `internal/diagfmt` | 4308 |
-| 7 | `internal/mir` | 4105 |
+| 7 | `internal/mir` | 4205 |
 | 8 | `internal/symbols` | 3952 |
 | 9 | `internal/ast` | 3785 |
 | 10 | `internal/mono` | 3391 |
@@ -36,8 +36,8 @@
 
 ## 📈 Total volume (code + tests)
 
-- **Files:** 481
-- **Lines of code:** 101060
+- **Files:** 489
+- **Lines of code:** 101305
 
 ## 📊 Percentage breakdown
