+++ internal/hir/lower_expr.go
@@ -1,11 +1,8 @@
 package hir
 
 import (
-	"strings"
-
 	"surge/internal/ast"
 	"surge/internal/sema"
-	"surge/internal/source"
 	"surge/internal/symbols"
 	"surge/internal/types"
 )
@@ -126,37 +123,7 @@ func (l *lowerer) lowerExprCore(exprID ast.ExprID) *Expr {
 	}
 }
 
-func (l *lowerer) intrinsicSymbolID(name string) symbols.SymbolID {
-	if l == nil || l.symRes == nil || l.symRes.Table == nil || !l.symRes.FileScope.IsValid() {
-		return symbols.NoSymbolID
-	}
-	nameID := l.strings.Intern(name)
-	return l.symbolInScope(l.symRes.FileScope, nameID, symbols.SymbolFunction)
-}
-
-func (l *lowerer) intrinsicCallee(name string, span source.Span) (*Expr, symbols.SymbolID) {
-	symID := l.intrinsicSymbolID(name)
-	return &Expr{
-		Kind: ExprVarRef,
-		Type: types.NoTypeID,
-		Span: span,
-		Data: VarRefData{Name: name, SymbolID: symID},
-	}, symID
-}
-
-func (l *lowerer) boolLiteralExpr(span source.Span, value bool) *Expr {
-	boolType := types.NoTypeID
-	if l != nil && l.module != nil && l.module.TypeInterner != nil {
-		boolType = l.module.TypeInterner.Builtins().Bool
-	}
-	return &Expr{
-		Kind: ExprLiteral,
-		Type: boolType,
-		Span: span,
-		Data: LiteralData{Kind: LiteralBool, BoolValue: value},
-	}
-}
-
+// lowerRangeLitExpr lowers a range literal expression.
 func (l *lowerer) lowerRangeLitExpr(expr *ast.Expr, ty types.TypeID) *Expr {
 	rangeData := l.builder.Exprs.RangeLits.Get(uint32(expr.Payload))
 	if rangeData == nil {
@@ -621,234 +588,5 @@ func (l *lowerer) lowerBlockExpr(expr *ast.Expr, ty types.TypeID) *Expr {
 		Type: ty,
 		Span: expr.Span,
 		Data: BlockExprData{Block: block},
-	}
-}
-
-// Helper functions
-
-func (l *lowerer) lookupString(id source.StringID) string {
-	if l.strings == nil || id == source.NoStringID {
-		return ""
-	}
-	s, _ := l.strings.Lookup(id)
-	return s
-}
-
-func (l *lowerer) varRefForSymbol(symID symbols.SymbolID, span source.Span) *Expr {
-	if !symID.IsValid() {
-		return nil
-	}
-	name := ""
-	ty := types.NoTypeID
-	if l.symRes != nil && l.symRes.Table != nil && l.symRes.Table.Symbols != nil {
-		if sym := l.symRes.Table.Symbols.Get(symID); sym != nil {
-			if sym.Name != source.NoStringID {
-				name = l.lookupString(sym.Name)
-			}
-			if sym.Type != types.NoTypeID {
-				ty = sym.Type
-			}
-		}
-	}
-	return &Expr{
-		Kind: ExprVarRef,
-		Type: ty,
-		Span: span,
-		Data: VarRefData{
-			Name:     name,
-			SymbolID: symID,
-		},
-	}
-}
-
-func (l *lowerer) applySelfBorrow(symID symbols.SymbolID, recv *Expr) *Expr {
-	if recv == nil || !symID.IsValid() {
-		return recv
-	}
-	if l.symRes == nil || l.symRes.Table == nil || l.symRes.Table.Symbols == nil {
-		return recv
-	}
-	sym := l.symRes.Table.Symbols.Get(symID)
-	if sym == nil || sym.Signature == nil || !sym.Signature.HasSelf || len(sym.Signature.Params) == 0 {
-		return recv
-	}
-	selfKey := string(sym.Signature.Params[0])
-	mut := false
-	switch {
-	case strings.HasPrefix(selfKey, "&mut "):
-		mut = true
-	case strings.HasPrefix(selfKey, "&"):
-	default:
-		return recv
-	}
-	if l.isReferenceType(recv.Type) {
-		return recv
-	}
-	refType := l.referenceType(recv.Type, mut)
-	op := ast.ExprUnaryRef
-	if mut {
-		op = ast.ExprUnaryRefMut
-	}
-	return &Expr{
-		Kind: ExprUnaryOp,
-		Type: refType,
-		Span: recv.Span,
-		Data: UnaryOpData{
-			Op:      op,
-			Operand: recv,
-		},
-	}
-}
-
-func (l *lowerer) isReferenceType(id types.TypeID) bool {
-	if id == types.NoTypeID || l.semaRes == nil || l.semaRes.TypeInterner == nil {
-		return false
-	}
-	tt, ok := l.semaRes.TypeInterner.Lookup(id)
-	if !ok {
-		return false
-	}
-	return tt.Kind == types.KindReference
-}
-
-func (l *lowerer) referenceType(elem types.TypeID, mutable bool) types.TypeID {
-	if elem == types.NoTypeID || l.semaRes == nil || l.semaRes.TypeInterner == nil {
-		return types.NoTypeID
-	}
-	return l.semaRes.TypeInterner.Intern(types.MakeReference(elem, mutable))
-}
-
-func (l *lowerer) lookupTypeFromAST(_ ast.TypeID) types.TypeID {
-	// AST TypeID is different from types.TypeID
-	// For now, we return NoTypeID - the actual type would need
-	// to be resolved through the type checker's type expressions
-	// This is a simplification - in a full implementation,
-	// we'd need to map AST type expressions to resolved types
-	return types.NoTypeID
-}
-
-func (l *lowerer) inferOwnership(ty types.TypeID) Ownership {
-	if ty == types.NoTypeID || l.semaRes == nil || l.semaRes.TypeInterner == nil {
-		return OwnershipNone
-	}
-
-	t, ok := l.semaRes.TypeInterner.Lookup(ty)
-	if !ok {
-		return OwnershipNone
-	}
-
-	switch t.Kind {
-	case types.KindReference:
-		if t.Mutable {
-			return OwnershipRefMut
-		}
-		return OwnershipRef
-	case types.KindPointer:
-		return OwnershipPtr
-	case types.KindOwn:
-		return OwnershipOwn
-	case types.KindInt, types.KindUint, types.KindFloat, types.KindBool:
-		return OwnershipCopy
-	default:
-		return OwnershipNone
-	}
-}
-
-//nolint:gocritic // ifElseChain is clearer than switch for character ranges
-func parseIntLiteral(s string) int64 {
-	var result int64
-	for _, c := range s {
-		if c >= '0' && c <= '9' {
-			result = result*10 + int64(c-'0')
-		} else if c == '_' {
-			continue
-		} else {
-			break
-		}
-	}
-	return result
-}
-
-//nolint:gocritic // ifElseChain is clearer than switch for character ranges
-func parseFloatLiteral(s string) float64 {
-	var result float64
-	var frac float64 = 1
-	var inFrac bool
-	for _, c := range s {
-		if c >= '0' && c <= '9' {
-			if inFrac {
-				frac /= 10
-				result += float64(c-'0') * frac
-			} else {
-				result = result*10 + float64(c-'0')
-			}
-		} else if c == '.' {
-			inFrac = true
-		} else if c == '_' {
-			continue
-		} else {
-			break
-		}
-	}
-	return result
-}
-
-// wrapInSome wraps an expression in a Some() tag constructor call.
-// This is used for implicit tag injection: let x: int? = 1 becomes Some(1).
-func (l *lowerer) wrapInSome(inner *Expr, targetType types.TypeID) *Expr {
-	return l.wrapInTagConstructor(inner, targetType, "Some")
-}
-
-// wrapInSuccess wraps an expression in a Success() tag constructor call.
-// This is used for implicit tag injection: let x: int! = 1 becomes Success(1).
-func (l *lowerer) wrapInSuccess(inner *Expr, targetType types.TypeID) *Expr {
-	return l.wrapInTagConstructor(inner, targetType, "Success")
-}
-
-// wrapInTagConstructor creates a call to a tag constructor wrapping the inner expression.
-func (l *lowerer) wrapInTagConstructor(inner *Expr, targetType types.TypeID, tagName string) *Expr {
-	if inner == nil {
-		return nil
-	}
-
-	// Look up the tag symbol for the constructor
-	var tagSymID symbols.SymbolID
-	if l.symRes != nil && l.symRes.Table != nil && l.strings != nil {
-		nameID := l.strings.Intern(tagName)
-		// Look up tag in file scope
-		if l.symRes.Table.Scopes != nil && l.symRes.FileScope.IsValid() {
-			scopeData := l.symRes.Table.Scopes.Get(l.symRes.FileScope)
-			if scopeData != nil {
-				if ids := scopeData.NameIndex[nameID]; len(ids) > 0 {
-					for _, id := range ids {
-						sym := l.symRes.Table.Symbols.Get(id)
-						if sym != nil && sym.Kind == symbols.SymbolTag {
-							tagSymID = id
-							break
-						}
-					}
-				}
-			}
-		}
-	}
-
-	// Create a call expression that wraps the inner expression
-	return &Expr{
-		Kind: ExprCall,
-		Type: targetType,
-		Span: inner.Span,
-		Data: CallData{
-			Callee: &Expr{
-				Kind: ExprVarRef,
-				Type: types.NoTypeID, // Callee type doesn't matter for dispatch
-				Span: inner.Span,
-				Data: VarRefData{
-					Name:     tagName,
-					SymbolID: tagSymID,
-				},
-			},
-			Args:     []*Expr{inner},
-			SymbolID: tagSymID,
-		},
 	}
 }
